# openapi.yaml
openapi: 3.0.3
info:
  title: Warden API
  description: |
    一个高性能的允许列表（AllowList）用户数据服务，支持本地和远程配置源的数据同步与合并。
    
    ## 核心特性
    - 高性能：支持每秒 5000+ 请求
    - 多数据源：支持本地配置文件和远程 API 两种数据源
    - 灵活策略：提供 6 种数据合并模式
    - 定时更新：基于 Redis 分布式锁的定时任务，自动同步数据
    - 容器化部署：完整的 Docker 支持
  version: 1.0.0
  contact:
    name: Warden Project
    url: https://github.com/soulteary/warden
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081
    description: 本地开发服务器
  - url: https://api.example.com
    description: 生产环境服务器
...
paths:
...
  /user:
    get:
      tags:
        - users
      summary: 查询单个用户
      description: |
        根据标识符查询单个用户信息。支持通过手机号、邮箱或用户ID查询。

        注意：只能提供一个查询参数（phone、mail 或 user_id）。
      operationId: getUserByIdentifier
      parameters:
        - name: phone
          in: query
          description: 用户手机号
          required: false
          schema:
            type: string
            example: "13800138000"
        - name: mail
          in: query
          description: 用户邮箱
          required: false
          schema:
            type: string
            format: email
            example: "admin@example.com"
        - name: user_id
          in: query
          description: 用户唯一标识符
          required: false
          schema:
            type: string
            example: "a1b2c3d4e5f6g7h8"
      responses:
        '200':
          description: 成功返回用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowListUser'
              example:
                phone: "13800138000"
                mail: "admin@example.com"
                user_id: "a1b2c3d4e5f6g7h8"
                status: "active"
                scope: ["read", "write"]
                role: "admin"
        '400':
          description: 请求参数错误（缺少标识符或提供了多个标识符）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Bad Request: missing identifier (phone, mail, or user_id)"
        '404':
          description: 用户未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "User not found"
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/lookup:
    get:
      tags:
        - users
      summary: 用户查询（Stargate/Herald 集成）
      description: |
        根据单一 identifier 查询用户，返回 user_id、destination（email/phone）、status、channel_hint，
        便于 Stargate 选择 Herald 的 sms/email 通道。
        identifier 自动识别：含 @ 按邮箱查，否则先按手机再按 user_id。
      operationId: getLookup
      parameters:
        - name: identifier
          in: query
          required: true
          schema:
            type: string
          description: 手机号、邮箱或 user_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LookupResponse'
        '400':
          description: 缺少 identifier 或参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 用户未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    AllowListUser:
      type: object
      description: 至少提供 phone 或 mail 之一；支持仅邮箱用户（phone 可为空）
      properties:
        phone:
          type: string
          description: 用户手机号（与 mail 二选一或同时提供；支持仅邮箱用户时为空）
          example: "13800138000"
        mail:
          type: string
          format: email
          description: 用户邮箱（与 phone 二选一或同时提供）
          example: "admin@example.com"
        user_id:
          type: string
          description: 用户唯一标识符（可选，如果未提供则自动生成）
          example: "a1b2c3d4e5f6g7h8"
        status:
          type: string
          description: 用户状态（如 "active", "inactive", "suspended"）
          enum:
            - active
            - inactive
            - suspended
          default: "active"
          example: "active"
        scope:
          type: array
          description: 用户权限范围（可选）
          items:
            type: string
          example: ["read", "write"]
        role:
          type: string
          description: 用户角色（可选）
          example: "admin"
    LookupResponse:
      type: object
      description: GET /v1/lookup 响应（Stargate/Herald 集成）
      properties:
        user_id:
          type: string
          description: 用户唯一标识
        destination:
          type: object
          properties:
            email:
              type: string
              description: 邮箱（可选）
            phone:
              type: string
              description: 手机号（可选）
        status:
          type: string
          description: 用户状态
        channel_hint:
          type: string
          enum: ["sms", "email"]
          description: OTP 推荐通道
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: 错误消息
          example: "Invalid pagination parameters"
  responses:
...
    MethodNotAllowed:
      description: 不支持的 HTTP 方法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Method not allowed"
    InternalServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"